move chat streaming model out of file
